api/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
opencv-python-headless
tensorflow
numpy
pillow
pillow-heif
matplotlib